BLD: add deploy target, archive for html packages, optional upload html for the correct runner, changed coloring mode

+++ .workflow.sh
@@ -5,20 +5,24 @@
 # leave any virtual environment
 deactivate 2>/dev/null || true
 
-function BLUE(){
-    echo -e '\033[0;34;49m'
+function green(){
+    # green message
+    echo -e '\033[0;32m'$1'\033[0m'
 }
-function GREEN(){
-    echo -e '\033[0;32m'
+
+function blue(){
+    # blue message
+    echo -e '\033[0;34;49m'$1'\033[0m'
 }
-function YELLOW(){
-    echo -e '\033[0;33;49m'
+
+function yellow(){
+    # yellow message
+    echo -e '\033[0;33;49m'$1'\033[0m'
 }
-function RED(){
-    echo -e '\033[0;31;49m'
-}
-function NCOL(){
-    echo -e '\033[0m'
+
+function red(){
+    # red message
+    echo -e '\033[0;31;49m'$1'\033[0m'
 }
 
 if [[ $- == *i* ]]; then
@@ -36,19 +40,16 @@ else
 fi
 
 if [ ! -z "$VIRTUAL_ENV" ]; then
-    RED
-    echo "active virtual environment $VIRTUAL_ENV : leave first"
-    echo "call deactivate"
-    NCOL
+    red "Active virtual environment $VIRTUAL_ENV - leave first"
+    red "Please call 'deactivate'"
     return
 fi
 
 function use_venv(){
     local venv_path=$1
     if [ -d $venv_path ]; then
-        GREEN
-        echo "Activating virtual environment $venv_path"
-        NCOL
+        green "Activating virtual environment $venv_path"
+
         if [ $OS == "Windows" ] || [ $OS == "Windows_NT" ]; then
             # windows
             source $venv_path/Scripts/activate
@@ -57,9 +58,7 @@ function use_venv(){
             source $venv_path/bin/activate
         fi
     else
-        GREEN
-        echo "Virtual environment $venv_path does not exist. Creating it."
-        NCOL
+        green "Virtual environment $venv_path does not exist. Creating it."
         new_venv $venv_path
     fi
 }
@@ -67,9 +66,7 @@ function use_venv(){
 
 function new_venv(){
     local venv_path=$1
-    GREEN
-    echo "Creating fresh virtual environment $venv_path"
-    NCOL
+    green "Creating fresh virtual environment $venv_path"
     $BASEPYTHON --version
     deactivate 2>/dev/null || true
 
@@ -85,15 +82,11 @@ function new_venv(){
 
 function testReport(){
     if [ -n "$VIRTUAL_ENV" ]; then
-        GREEN
-        echo "Testing build in active virtual environment: $VIRTUAL_ENV"
-        echo "Python: $(python -V 2>&1) ($(python -c 'import sys; print(sys.executable)'))"
-        NCOL
+        green "Testing build in active virtual environment: $VIRTUAL_ENV"
+        green "Python: $(python -V 2>&1) ($(python -c 'import sys; print(sys.executable)'))"
     else
-        YELLOW
-        echo "No active virtual environment."
-        echo "Python: $(python -V 2>&1) ($(python -c 'import sys; print(sys.executable)'))"
-        NCOL
+        yellow "No active virtual environment."
+        yellow "Python: $(python -V 2>&1) ($(python -c 'import sys; print(sys.executable)'))"
     fi
 
     python -c 'import pygimli; print(pygimli.version())'
@@ -102,9 +95,7 @@ function testReport(){
 
 
 function clean(){
-    GREEN
-    echo "*** Cleaning ...                 ***"
-    NCOL
+    green "*** clean  (Remove old build artifacts) ***"
 
     pushd $PROJECT_ROOT
         echo "clear pygimli cache"
@@ -129,10 +120,7 @@ function clean(){
 
 
 function build_pre(){
-    GREEN
-    echo "*** build_pre (Prepare building environment) ***"
-    NCOL
-
+    green "*** build_pre (Prepare building environment) ***"
     ## clean previous build, dist, and doc artifacts
     clean
 
@@ -148,14 +136,10 @@ function build_pre(){
 
 
 function build(){
-    GREEN
-    echo "*** build (Building now ...) ***"
-    NCOL
+    green "*** build (Building now ...) ***"
 
     if [ ! -d $BUILD_DIR ]; then
-        YELLOW
-        echo "BUILD_DIR: $BUILD_DIR not found. Preparing now (build_pre)."
-        NCOL
+        yellow "BUILD_DIR: $BUILD_DIR not found. Preparing now (build_pre)."
         build_pre
     fi
 
@@ -195,9 +179,7 @@ function build(){
 
 
 function build_whls(){
-    GREEN
-    echo "*** build_wheels (Building wheels now ...) ***"
-    NCOL
+    green "*** build_wheels (Building wheels now ...) ***"
 
     if [ "$OS" == "Windows" ] || [ "$OS" == "Windows_NT" ]; then
         LIBGIMLI=$(ls $BUILD_DIR/bin/libgimli.dll | head -n 1)
@@ -211,14 +193,10 @@ function build_whls(){
     fi
 
     if [ -z $LIBGIMLI ]; then
-        YELLOW
-        echo "libgimli not found in $BUILD_DIR/lib. Building now (build)."
-        NCOL
+        yellow "libgimli not found in $BUILD_DIR/lib. Building now (build)."
         build
     else
-        GREEN
-        echo "libgimli found: $LIBGIMLI."
-        NCOL
+        green "libgimli found: $LIBGIMLI."
     fi
 
     pushd $PROJECT_ROOT
@@ -239,29 +217,21 @@ function build_whls(){
                 WHLFILE=$(ls $WHEELHOUSE/pgcore*.whl | head -n 1)
 
                 if [ "$OS" == "MacOS" ] || [ "$(uname -s)" == "Darwin" ]; then
-                    BLUE
-                    echo "Repairing pgcore whl for MacOS ($WHLFILE)"
-                    NCOL
+                    blue "Repairing pgcore whl for MacOS ($WHLFILE)"
                     delocate-wheel -v $WHLFILE
 
                 elif [ "$OS" == "Windows" ] || [ "$OS" == "Windows_NT" ]; then
-                    BLUE
-                    echo "Repairing pgcore whl for Windows ($WHLFILE)"
-                    NCOL
+                    blue "Repairing pgcore whl for Windows ($WHLFILE)"
                     delvewheel repair $WHLFILE --add-path $BUILD_DIR/bin/
 
                 else
-                    BLUE
-                    echo "Repairing pgcore whl for Linux ($WHLFILE)"
-                    NCOL
+                    blue "Repairing pgcore whl for Linux ($WHLFILE)"
                     echo "Build pgcore whl for Linux"
                 fi
             popd
 
-            GREEN
-            echo "Copying pgcore whl ($WHLFILE) to build dist $BUILD_DIR/dist/"
+            green "Copying pgcore whl ($WHLFILE) to build dist $BUILD_DIR/dist/"
             cp $WHLFILE $BUILD_DIR/dist/
-            NCOL
         popd
 
         ### create pygimli wheel
@@ -274,9 +244,7 @@ function build_whls(){
 
 function install_WHL_E(){
     opt=$1
-    GREEN
-    echo "*** install pygimli $opt from whl files (editable) ***"
-    NCOL
+    green "*** install pygimli $opt from whl files (editable) ***"
 
     pushd $PROJECT_ROOT
         # special case for windows .. pgcore install to ensuse mingw runtime libs are found
@@ -293,9 +261,8 @@ function install_WHL_E(){
 
 function install_WHL(){
     opt=$1
-    GREEN
-    echo "*** install pygimli $opt from whl files (non editable) ***"
-    NCOL
+    green "*** install pygimli $opt from whl files (non editable) ***"
+
     pushd $PROJECT_ROOT
         uv pip uninstall pygimli pgcore
         uv pip install --force-reinstall $PROJECT_DIST/pgcore*.whl
@@ -307,32 +274,25 @@ function install_WHL(){
 
 
 function build_post(){
-    GREEN
-    echo "*** build_post (Testing build) ***"
-    NCOL
+    green "*** build_post (Testing build) ***"
 
     PGCORE_WHL=$(ls $BUILD_DIR/dist/pgcore*.whl | head -n 1)
     PG_WHL=$(ls $BUILD_DIR/dist/pygimli*.whl | head -n 1)
 
     if [ -z $PGCORE_WHL ] || [ -z $PG_WHL ]; then
-        YELLOW
-        echo "pgcore or pygimli whl's not found in $BUILD_DIR/dist. Building now (build_whls)"
-        NCOL
+        yellow "pgcore or pygimli whl's not found in $BUILD_DIR/dist. Building now (build_whls)"
         build_whls
     else
-        GREEN
-        echo "Whl's found in build dist:"
-        echo -e "\t pgcore: $PGCORE_WHL"
-        echo -e "\t pygimli: $PG_WHL"
-        NCOL
+        green "Whl's found in build dist:"
+        green "\t pgcore: $PGCORE_WHL"
+        green "\t pygimli: $PG_WHL"
     fi
 
     pushd $PROJECT_ROOT
         mkdir -p $PROJECT_DIST
 
-        GREEN
-        echo "Copying built whl files to project dist: $PROJECT_DIST"
-        NCOL
+        green "Copying built whl files to project dist: $PROJECT_DIST"
+
         cp $BUILD_DIR/dist/pgcore*.whl $PROJECT_DIST/
         cp $BUILD_DIR/dist/pygimli*.whl $PROJECT_DIST/
 
@@ -350,17 +310,17 @@ function build_post(){
 }
 
 
-function test_pre(){
-    GREEN
-    echo "*** test_pre (Prepare testing environment) ***"
-    NCOL
-
+function need_build_post(){
     if [ ! -f $PROJECT_DIST/pgcore*.whl ]; then
-        YELLOW
-        echo "pgcore wheel not found in project dist. Building first. (build_post)"
-        NCOL
+        yellow "pgcore wheel not found in project dist. Building first. (build_post)"
         build_post
     fi
+}
+
+
+function test_pre(){
+    green "*** test_pre (Prepare testing environment) ***"
+    need_build_post
 
     pushd $PROJECT_ROOT
         new_venv $VENV_TEST
@@ -370,9 +330,7 @@ function test_pre(){
 
 
 function test(){
-    GREEN
-    echo "*** test (Testing now ...) ***"
-    NCOL
+    green "*** test (Testing now ...) ***"
 
     pushd $PROJECT_ROOT
         if [ ! -d $VENV_TEST ]; then
@@ -388,16 +346,8 @@ function test(){
 
 
 function doc_pre(){
-    GREEN
-    echo "*** doc_pre (Preparing documentation ...) ***"
-    NCOL
-
-    if [ ! -f $PROJECT_DIST/pgcore*.whl ]; then
-        YELLOW
-        echo "pgcore wheel not found in project dist. Building first. (build_post)"
-        NCOL
-        build_post
-    fi
+    green "*** doc_pre (Preparing documentation ...) ***"
+    need_build_post
 
     pushd $PROJECT_ROOT
         new_venv $VENV_DOC
@@ -407,19 +357,13 @@ function doc_pre(){
 
 
 function doc(){
-    GREEN
-    echo "*** doc (Creating documentation) ***"
-    NCOL
+    green "*** doc (Creating documentation) ***"
 
     use_venv $VENV_DOC
     if python -c 'import sphinx' &>/dev/null; then
-        GREEN
-        echo "sphinx is installed"
-        NCOL
+        green "sphinx is installed"
     else
-        YELLOW
-        echo "sphinx is NOT installed (calling doc_pre)"
-        NCOL
+        yellow "sphinx is NOT installed (calling doc_pre)"
         doc_pre
     fi
 
@@ -445,25 +389,30 @@ function doc(){
 
 
 function doc_post(){
-    GREEN
-    echo "*** doc_post (Deploying html) ***"
-    NCOL
+    green "*** doc_post (Deploying html) ***"
 
     if [ ! -f $BUILD_DIR/doc/_build/html/index.html ]; then
-        YELLOW
-        echo "Documentation html not found in dist. Building first. (doc)"
-        NCOL
+        yellow "Documentation html not found in dist. Building first. (doc)"
         doc
     fi
 
-    # Should only run on pgserver
-    rsync -avP --delete $BUILD_DIR/doc/_build/html/ /var/www/html
+    use_venv $VENV_DOC
+    VERSION=$(python -c 'import pygimli; print(pygimli.__version__.split("(")[0].strip())')
+    green "Deploying documentation for pygimli version: $VERSION"
 
-    #pushd $BUILD_DIR/doc/_build
-        #cp -r html $PROJECT_DIST/html
-        #tar -czvf $PROJECT_DIST/html.tgz html
-    #popd
+    if [ "$RUNNER_NAME" == "pgserver" ]; then
+        # Should only run on pgserver
+        green "On pgserver - deploying documentation to /var/www/html"
+        rsync -avP --delete $BUILD_DIR/doc/_build/html/ /var/www/html
+    else
+        yellow "Not on pgserver - skipping documentation deployment"
+    fi
 
+    mkdir -p $ARCHIVE_DIR
+    pushd $BUILD_DIR/doc/_build
+        tar -czf $ARCHIVE_DIR/html-$VERSION.tgz html
+    popd
+    green "Archived documentation to $ARCHIVE_DIR/html-$VERSION.tgz"
     #rsync -aP $DISTPATH/html user@pygimli.org:DEV_HTML_PATH --delete
 
     # SNAP_PATH=`date +"%Y%m%d"`
@@ -476,17 +425,10 @@ function doc_post(){
     # python -m pip install --force ~/snapshots/latest/pygimli*
 }
 
-function install(){
-    GREEN
-    echo "*** Install (Creating editable installation in build venv) ***"
-    NCOL
 
-    if [ ! -f $PROJECT_DIST/pgcore*.whl ]; then
-        YELLOW
-        echo "pgcore wheel not found in project dist. Building first.(build_post)"
-        NCOL
-        build_post
-    fi
+function install(){
+    green "*** Install (Creating editable installation in build venv) ***"
+    need_build_post
 
     pushd $PROJECT_ROOT
         use_venv $VENV_PYGIMLI
@@ -494,34 +436,70 @@ function install(){
         testReport
     popd
 
-    GREEN
-    echo "Editable installation created in venv: $VENV_PYGIMLI"
-    echo ""
-    echo "To use it, call: "
-    echo ""
-    echo "source $VENV_PYGIMLI/bin/activate #(linux/macos)"
-    echo "source $VENV_PYGIMLI/Scripts/activate #(windows)"
-    NCOL
+    green "Editable installation created in venv: $VENV_PYGIMLI"
+    green ""
+    green "To use it, call: "
+    green ""
+    green "source $VENV_PYGIMLI/bin/activate #(linux/macos)"
+    green "source $VENV_PYGIMLI/Scripts/activate #(windows)"
 }
+
+
+function twine(){
+    WHL_FILE=$1
+    green "*** twine upload whl file: $WHL_FILE ***"
+    python -m twine upload --repository pypi $WHL_FILE
+}
+
+
+function deploy(){
+    target_whl=$1
+    if [ -z $target_whl ]; then
+        red "give target_whl=\"pygimli\" or \"pgcore\" as argument"
+        return
+    fi
+    green "*** Deploy $target_whl file ***"
+    need_build_post
+    use_venv $VENV_BUILD
+
+    TARGET_WHL=$(ls $PROJECT_DIST/$target_whl*.whl | head -n 1)
+
+    if [ -z $TARGET_WHL ]; then
+        red "Target whl file $target_whl not found in $PROJECT_DIST"
+        return
+    else
+        blue "FOUND: $TARGET_WHL"
+    fi
+    echo "Do you wish to deploy ? (can't be undone)"
+    select yn in "Yes" "No"; do
+        case $yn in
+            Yes ) twine $TARGET_WHL; break;;
+            No ) return;;
+        esac
+    done
+    return
+}
+
 
 function help(){
     echo ""
     echo "run: ${BASH_SOURCE[0]} TARGET"
     echo ""
     echo "TARGETS:"
-    echo "    help       show this help"
-    echo "    clean      remove build artifacts for PYVERSION"
-    echo "    build_pre  prepare build environment venv"
-    echo "    build      [build_pre] build project"
-    echo "    build_whls [build] build whl files for pgcore and pygimli"
-    echo "    build_post [build] test build"
-    echo "    test_pre   [build] prepare test environment"
-    echo "    test       [test_pre] run tests"
-    echo "    doc_pre    [build] prepare documentation environment"
-    echo "    doc        [doc_pre] build documentation"
-    echo "    doc_post   [doc] deploy documentation"
-    echo "    install    [build] Create default editable installation in venv"
-    echo "    all        [clean build test doc]"
+    echo "    help           show this help"
+    echo "    clean          remove build artifacts for PYVERSION"
+    echo "    build_pre      prepare build environment venv"
+    echo "    build          [build_pre] build project"
+    echo "    build_whls     [build] build whl files for pgcore and pygimli"
+    echo "    build_post     [build] test build"
+    echo "    test_pre       [build] prepare test environment"
+    echo "    test           [test_pre] run tests"
+    echo "    doc_pre        [build] prepare documentation environment"
+    echo "    doc            [doc_pre] build documentation"
+    echo "    doc_post       [doc] deploy documentation"
+    echo "    install        [build] Create default editable installation in venv"
+    echo "    deploy         [build_post] deploy wheel files"
+    echo "    all            [clean build test doc]"
     echo ""
     echo "ENVIRONMENT variables:"
     echo "    BASEPYTHON   base python interpreter. Default system python3."
@@ -533,6 +511,8 @@ function help(){
     echo "    PYVERSION=3.12 bash ${BASH_SOURCE[0]} all"
     echo ""
 }
+
+
 function all(){
     clean
     build_post
@@ -543,20 +523,19 @@ function all(){
 JOB_NUM=0
 
 if [ ! -z $GITHUB_ACTIONS ]; then
-    GREEN
-    echo "GITHUB action runner on WORKSPACE=$WORKSPACE at $RUNNER_NAME"
-    echo "RUNNER_ARCH=$RUNNER_ARCH"
-    echo "RUNNER_TEMP=$RUNNER_TEMP"
-    echo "GITHUB_REF_NAME=$GITHUB_REF_NAME"
-    echo "GITHUB_JOB=$GITHUB_JOB"
-    echo "GITHUB_REPOSITORY=$GITHUB_REPOSITORY"
-    echo "GITHUB_WORKSPACE=$GITHUB_WORKSPACE"
-    NCOL
+    green "GITHUB action runner on WORKSPACE=$WORKSPACE at $RUNNER_NAME"
+    green "RUNNER_ARCH=$RUNNER_ARCH"
+    green "RUNNER_TEMP=$RUNNER_TEMP"
+    green "GITHUB_REF_NAME=$GITHUB_REF_NAME"
+    green "GITHUB_JOB=$GITHUB_JOB"
+    green "GITHUB_REPOSITORY=$GITHUB_REPOSITORY"
+    green "GITHUB_WORKSPACE=$GITHUB_WORKSPACE"
 
     WORKSPACE=$GITHUB_WORKSPACE
     JOB_NUM=$GITHUB_RUN_NUMBER
     OS=$RUNNER_OS
     SYSTEM=$(uname -s)
+
 elif [ -z $WORKSPACE ]; then
     WORKSPACE=$(realpath $(pwd))
     if [ -z $OS ]; then
@@ -574,13 +553,9 @@ elif [ -z $WORKSPACE ]; then
         esac
     fi
     SYSTEM=$(uname -s)
-    GREEN
-    echo "Local Build (no Jenkins) on WORKSPACE=$WORKSPACE"
-    NCOL
+    green "Local Build (no Jenkins) on WORKSPACE=$WORKSPACE"
 else
-    GREEN
-    echo "Unknown CI Build on WORKSPACE=$WORKSPACE"
-    NCOL
+    green "Unknown CI Build on WORKSPACE=$WORKSPACE"
     WORKSPACE=$(realpath $(pwd))
 fi
 
@@ -595,9 +570,7 @@ if [ -z $SOURCE_DIR ]; then
     done
     if [ "$CAND" = "/" ]; then
         # fallback: use the immediate directory name of the script
-        RED
-        echo "Could not find project root marker (.git or pyproject.toml). Using script directory name as SOURCE_DIR."
-        NCOL
+        red "Could not find project root marker (.git or pyproject.toml). Using script directory name as SOURCE_DIR."
     else
         SOURCE_DIR=$(basename "$CAND")
     fi
@@ -617,10 +590,8 @@ function abspath() {
 
     # Fallback to python if realpath not present or failed
     if [ -z "$abs" ]; then
-        RED
-        echo "Could not determine absolute path for $p"
+        red "Could not determine absolute path for $p"
         return
-        NCOL
     fi
 
     # Normalize empty result
@@ -645,10 +616,8 @@ if [ -z $BASEPYTHON ]; then
     if [ -z $PYVERSION ]; then
 
         if [ ! -x "$(command -v python3)" ]; then
-            RED
-            echo "python3 not found in PATH. Please install python3 or set BASEPYTHON to a valid python interpreter."
-            echo "e.g., BASEPYTHON=../../miniconda3/python bash ${BASH_SOURCE[0]}"
-            NCOL
+            red "python3 not found in PATH. Please install python3 or set BASEPYTHON to a valid python interpreter."
+            red "e.g., BASEPYTHON=../../miniconda3/python bash ${BASH_SOURCE[0]}"
             return
         else
             echo "Using system python3 as BASEPYTHON"
@@ -679,6 +648,7 @@ VENV_TEST=$(realpath $WORKSPACE/venv-test-py$PYVERSION)
 VENV_DOC=$(realpath $WORKSPACE/venv-doc-py$PYVERSION)
 VENV_PYGIMLI=$(realpath $WORKSPACE/venv-pygimli-py$PYVERSION)
 PROJECT_DIST=$(realpath $WORKSPACE/dist-py$PYVERSION)
+ARCHIVE_DIR=$(realpath $WORKSPACE/archive)
 
 echo "JOB_NUM=$JOB_NUM"
 echo "WORKSPACE=$WORKSPACE"
@@ -725,6 +695,8 @@ do
         doc_post;;
     install)
         install;;
+    deploy)
+        deploy $2;;
     all)
         all;;
     help)
